More changes

+++ src/main/java/org/getspout/spout/inventory/SimpleMaterialManager.java
@@ -27,9 +27,7 @@ import java.util.List;
 import gnu.trove.list.array.TByteArrayList;
 import gnu.trove.list.array.TIntArrayList;
 import gnu.trove.map.hash.TIntObjectHashMap;
-
 import net.minecraft.server.v1_6_R3.Item;
-
 import org.bukkit.Bukkit;
 import org.bukkit.World;
 import org.bukkit.block.Block;
@@ -37,7 +35,6 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.Recipe;
 import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.java.JavaPlugin;
-
 import org.getspout.spout.block.SpoutCraftBlock;
 import org.getspout.spout.player.SpoutCraftPlayer;
 import org.getspout.spoutapi.SpoutManager;
@@ -64,7 +61,7 @@ public class SimpleMaterialManager extends AbstractBlockManager implements Mater
 
 	public static void disableFlintStackMix() {
 		try {
-			Method a = Item.class.getDeclaredMethod("a", new Class[] { boolean.class });
+			Method a = Item.class.getDeclaredMethod("a", new Class[] {boolean.class});
 			a.setAccessible(true);
 			a.invoke(Item.byId[318], Boolean.TRUE);
 		} catch (Exception e) {
@@ -83,13 +80,13 @@ public class SimpleMaterialManager extends AbstractBlockManager implements Mater
 		if (player.isSpoutCraftEnabled()) {
 			for (CustomBlock block : MaterialData.getCustomBlocks()) {
 				if (block instanceof SpoutPacket) {
-					player.sendPacket((SpoutPacket)block);
+					player.sendPacket((SpoutPacket) block);
 				}
 			}
 			for (CustomItem item : MaterialData.getCustomItems()) {
 				CustomBlock owner = MaterialData.getCustomBlock(item.getCustomId());
 				if (item instanceof SpoutPacket && owner == null) {
-					player.sendPacket((SpoutPacket)item);
+					player.sendPacket((SpoutPacket) item);
 				}
 			}
 		}
@@ -150,7 +147,7 @@ public class SimpleMaterialManager extends AbstractBlockManager implements Mater
 		int blockId = customBlock.getCustomId();
 
 		SpoutManager.getChunkDataManager().setBlockData(blockIdString, world, x, y, z, blockId);
-		((SpoutChunk) world.getChunkAt(x<<4, z<<4)).setCustomBlockData(x, y, z, data);
+		((SpoutChunk) world.getChunkAt(x << 4, z << 4)).setCustomBlockData(x, y, z, data);
 
 		queueBlockOverrides(world, x, y, z, blockId, data);
 
@@ -212,10 +209,11 @@ public class SimpleMaterialManager extends AbstractBlockManager implements Mater
 	}
 
 	private boolean glassUpdated = false;
+
 	// Fired when MaterialData.addCustomItem or MaterialData.addCustomBlock is called
 	public void onCustomMaterialRegistered(Material mat) {
 		if (mat instanceof CustomBlock && !glassUpdated) {
-			if (!((CustomBlock)mat).isOpaque()) {
+			if (!((CustomBlock) mat).isOpaque()) {
 				org.getspout.spout.block.mcblock.CustomMCBlock.updateGlass();
 				glassUpdated = true;
 			}
@@ -229,6 +227,7 @@ public class SimpleMaterialManager extends AbstractBlockManager implements Mater
 		private TIntArrayList zCoords = new TIntArrayList();
 		private TIntArrayList typeIds = new TIntArrayList();
 		private TByteArrayList data = new TByteArrayList();
+
 		BlockOverrides(World world) {
 			this.world = world;
 		}
@@ -248,7 +247,7 @@ public class SimpleMaterialManager extends AbstractBlockManager implements Mater
 				if (xCoords.size() > 128) {
 					int chunkX = xCoords.get(0) >> 4;
 					int chunkZ = zCoords.get(0) >> 4;
-					packet = new PacketOverrideChunk(SpoutManager.getChunkDataManager().getCustomBlockIds(world, chunkX, chunkZ), SpoutManager.getChunkDataManager().getCustomBlockData(world, chunkX, chunkZ),chunkX, chunkZ);
+					packet = new PacketOverrideChunk(SpoutManager.getChunkDataManager().getCustomBlockIds(world, chunkX, chunkZ), SpoutManager.getChunkDataManager().getCustomBlockData(world, chunkX, chunkZ), chunkX, chunkZ);
 				} else {
 					packet = new PacketOverrideMultiBlock(xCoords, yCoords, zCoords, typeIds, data);
 				}
@@ -263,7 +262,7 @@ public class SimpleMaterialManager extends AbstractBlockManager implements Mater
 				}
 			} else {
 				for (int i = 0; i < xCoords.size(); i++) {
-					SpoutPacket packet = new PacketOverrideBlock(xCoords.get(i), yCoords.get(i), zCoords.get(i), typeIds.get(i), data.get(i));
+					SpoutPacket packet = new PacketOverrideBlock(xCoords.get(i), yCoords.get(i), zCoords.get(i), (short) typeIds.get(i), data.get(i));
 					for (Player player : players) {
 						if (player instanceof SpoutCraftPlayer) {
 							SpoutCraftPlayer spc = (SpoutCraftPlayer) player;
